organize button directory
